relocate list and pages de-serialization functions
